ci: extract SonarCloud configuration script to GitHub action
