+++ .github/actions/configure-sonar-report-paths/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "configure-sonar-report-paths",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "main": "index.js",
+  "engines": {
+    "node": "16.19.1",
+    "yarn": "Use the npm CLI instead of Yarn Classic"
+  },
+  "volta": {
+    "node": "16.19.1"
+  },
+  "dependencies": {
+    "@actions/core": "^1.10.0",
+    "glob": "8.0.3",
+    "replace-in-file": "6.3.5"
+  }
+}
